refactor(core/managed): Remove resource status popover to its own module
