+++ app/scripts/modules/core/src/managed/ManagedResourceStatusIndicator.less
@@ -25,8 +25,4 @@
     font-size: 14px;
     border-radius: 50%;
   }
-
-  i {
-    color: var(--color-white);
-  }
 }
